Moved net_relay_server from offst-relay to offst-bin.
