+++ components/bin/src/strelay/mod.rs
@@ -1,3 +1,4 @@
+mod net_relay;
 mod strelaylib;
 
 pub use self::strelaylib::{strelay, RelayServerBinError, StRelayCmd};
