+++ components/relay/Cargo.toml
@@ -13,8 +13,6 @@ identity = { path = "../identity", version = "0.1.0" , package = "offst-identity
 timer = { path = "../timer", version = "0.1.0" , package = "offst-timer" }
 proto = { path = "../proto", version = "0.1.0" , package = "offst-proto" }
 keepalive = { path = "../keepalive", version = "0.1.0" , package = "offst-keepalive" }
-version = { path = "../version", version = "0.1.0" , package = "offst-version" }
-secure-channel = { path = "../secure_channel", version = "0.1.0" , package = "offst-secure-channel" }
 
 log = "0.4"
 futures = "0.3.1"
