+++ components/relay/src/server/mod.rs
@@ -1,5 +1,9 @@
 mod conn_limiter;
 mod conn_processor;
-pub mod net_server;
+// pub mod net_server;
 mod server;
+mod server_loop;
 mod types;
+
+pub use server::relay_server;
+pub use server_loop::RelayServerError;
